Moved tlds to common constants
