+++ apps/chrome/src/manifest.json
@@ -5,7 +5,12 @@
   "version": "1.2.0",
   "description": "Filter domains from your Google searches.",
   "homepage_url": "https://github.com/KMCGamer/google-domain-filter",
-  "permissions": ["storage", "activeTab", "webRequest"],
+  "permissions": [
+    "storage",
+    "activeTab",
+    "webRequest",
+    "declarativeNetRequest"
+  ],
   "host_permissions": [],
   "action": {
     "default_popup": "./popup/index.html"
@@ -30,11 +35,5 @@
     "16": "./static/logo16.png",
     "48": "./static/logo48.png",
     "128": "./static/logo128.png"
-  },
-  "browser_specific_settings": {
-    "gecko": {
-      "id": "gsb@carhart.dev",
-      "strict_min_version": "57.0"
-    }
   }
 }
